feat: Implement localized date formatting and update orders functionality
This commit introduces a `toLocalizedDateTime()` extension function for Strings to format ISO date strings into localized date-time representations. This function handles both Arabic and default locales.

The `OrdersMapper` has been updated to utilize this new extension for the `orderDate` field. Additionally, it now correctly maps `arabicMedicineName` based on the current locale and uses `totalPriceAfterDiscount` for `medicinePrice` in `OrderDetailsDto`.

The `OrdersViewModel` now includes the `getPharmacyUseCase` in its constructor and has improved logging for debugging purposes.

Finally, the `discountPercentage` type in `OrderDetailsDto` has been changed from `Int` to `Double`.

+++ app/src/main/java/com/example/atonce/data/remote/dto/OrdersResponse.kt
@@ -27,7 +27,7 @@ data class OrderDto(
 data class OrderDetailsDto(
     val arabicMedicineName: String,
     val discountAmount: Double,
-    val discountPercentage: Int,
+    val discountPercentage: Double,
     val medicineImage: String,
     val medicineName: String,
     val medicinePrice: Double,
